+++ src/main/java/org/entur/gbfs/loader/GbfsFeed.java
@@ -1,0 +1,5 @@
+package org.entur.gbfs.loader;
+
+import java.net.URI;
+
+public record GbfsFeed<S, T>(S name, Class<T> implementingClass, URI uri) {}
