+++ src/main/java/org/entur/gbfs/loader/GbfsSubscription.java
@@ -1,7 +1,6 @@
-package org.entur.gbfs;
+package org.entur.gbfs.loader;
 
 public interface GbfsSubscription {
-
   void init();
 
   boolean getSetupComplete();
