+++ src/test/java/org/entur/gbfs/GbfsAuthenticationTest.java
@@ -6,8 +6,8 @@ import java.util.concurrent.CountDownLatch;
 import java.util.function.Consumer;
 import org.entur.gbfs.authentication.Oauth2ClientCredentialsGrantRequestAuthenticator;
 import org.entur.gbfs.authentication.RequestAuthenticator;
-import org.entur.gbfs.v2.GbfsV2Delivery;
-import org.entur.gbfs.v2.GbfsV2Loader;
+import org.entur.gbfs.loader.v2.GbfsV2Delivery;
+import org.entur.gbfs.loader.v2.GbfsV2Loader;
 import org.entur.gbfs.v2_3.free_bike_status.GBFSFreeBikeStatus;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Disabled;
